Use autoload instead of require_relative:
This project already did a reasonably good job of deferring loads until
necessary, but this improves the situation a bit more. Deferring loading
of the various Command classes is relatively significant.

Overall this saves about 5ms on most invocations:

    % hyperfine -w1 'cd gel-require && shadowenv exec -- gel help' 'cd gel-autoload && shadowenv exec -- gel help'
    Benchmark 1: cd gel-require && shadowenv exec -- gel help
      Time (mean ± σ):      33.7 ms ±   2.0 ms    [User: 17.8 ms, System: 10.0 ms]
      Range (min … max):    29.2 ms …  42.1 ms    75 runs

    Benchmark 2: cd gel-autoload && shadowenv exec -- gel help
      Time (mean ± σ):      28.2 ms ±   2.6 ms    [User: 14.2 ms, System: 8.5 ms]
      Range (min … max):    25.2 ms …  46.2 ms    88 runs

    Summary
      'cd gel-autoload && shadowenv exec -- gel help' ran
        1.19 ± 0.13 times faster than 'cd gel-require && shadowenv exec -- gel help'

(both of these branches are running ruby with --disable-gems to get more
confounders out of the way)